fix: pullRequest.payload is now immutable
This work is to fix the mutability nature of dictionary. Dictionary data struct is mutable that means even if you declare new variables to hold the original dictionary and you change the declared variable, it will affect the original data.

What happened is that I used to pass the `PULL_REQUST` to `helpers.raise_pull_request()` and it will process the `payload.description` internally by joining the list elements with `\n`. It turns out this affects the original `PULL_REQUEST['payload']['description']` and this is not desired as this will affect other repos pull request description.

I tried fixing this using `dict(PULL_REQUEST)` and `PULL_REQUEST.copy()` to do shallow copy but the result was the same. It turns out when you have nested dicitonaries shallow copying won't much help you. So I used `copy.deepcopy()` to implement deep copying.

#### References:
- [Stackoverflow thread][stack-overflow-thread].
- [Shallow copy vs deep copy][python-docs]

[stack-overflow-thread]: https://stackoverflow.com/questions/5105517/deep-copy-of-a-dict-in-python
[python-docs]: https://docs.python.org/3/library/copy.html#:~:text=A%20shallow%20copy%20constructs%20a%20new%20compound%20object%20and%20then%20(to%20the%20extent%20possible)%20inserts%20references%20into%20it%20to%20the%20objects%20found%20in%20the%20original.