feat: dev or prod mode only
Users can optionally set mode in their targets config file to either 'dev' for development usage or to 'prod' for production usage. Defaults to 'prod'.

dev mode: used for development for testing patterns matching and confirming the results. In that mode, Code Migration Assistant will only search for patterns without modifying the files content (replacing) and will not create a branch nor will it push the changes and raise a pull request

prod mode: This is used when the user is confident about his pattern matching and he is ready migrate his code. User can optionally choose not to add pullRequest field in the targets config file to ignore raising pull requests if required.