fix: helpers.git.checkout_branch()
This change enhance the checkout_branch function and fix it.

Previously, checkout_branch was unable to checkout to a branch that was created by a previous run and its local repo was deleted from remote-targets/. As when you clone a remote repo, You only can see refs/remotes/origin/* branches and only refs/heads/<default_branch>. checkout_branch used to look for the branch only in the refs/heads and hence was unable to identify and then it will create a refs/heads branch based on the from_branch. This was wrong.

Additionally, I used to think from_branch would work but I haven't tested it on any branch but the refs/heads/main and it was working but if I were to supply a different from_branch it wouldn't have worked because it was only going to search for the refs/heads/ branches.